Conflicts:
	BNK/android/reference_wallet/fermat-bnk-android-reference-wallet-bank-money-bitdubai/src/main/java/com/bitdubai/reference_wallet/bank_money_wallet/fragments/details/TransactionDetailFragment.java
	BNK/plugin/bank_money_transaction/fermat-bnk-plugin-bank-money-transaction-deposit-bitdubai/src/main/java/com/bitdubai/fermat_bnk_plugin/layer/bank_money_transaction/deposit/developer/bitdubai/version_1/structure/BankMoneyTransactionRecordImpl.java
	BNK/plugin/bank_money_transaction/fermat-bnk-plugin-bank-money-transaction-deposit-bitdubai/src/main/java/com/bitdubai/fermat_bnk_plugin/layer/bank_money_transaction/deposit/developer/bitdubai/version_1/structure/DepositBankMoneyTransactionManager.java
	BNK/plugin/bank_money_transaction/fermat-bnk-plugin-bank-money-transaction-hold-bitdubai/src/main/java/com/bitdubai/fermat_bnk_plugin/layer/bank_money_transaction/hold/developer/bitdubai/version_1/structure/BankMoneyTransactionRecordImpl.java
	BNK/plugin/bank_money_transaction/fermat-bnk-plugin-bank-money-transaction-hold-bitdubai/src/main/java/com/bitdubai/fermat_bnk_plugin/layer/bank_money_transaction/hold/developer/bitdubai/version_1/structure/HoldBankMoneyTransactionProcessorAgent.java
	BNK/plugin/bank_money_transaction/fermat-bnk-plugin-bank-money-transaction-unhold-bitdubai/src/main/java/com/bitdubai/fermat_bnk_plugin/layer/bank_money_transaction/unhold/developer/bitdubai/version_1/structure/BankMoneyTransactionRecordImpl.java
	BNK/plugin/bank_money_transaction/fermat-bnk-plugin-bank-money-transaction-unhold-bitdubai/src/main/java/com/bitdubai/fermat_bnk_plugin/layer/bank_money_transaction/unhold/developer/bitdubai/version_1/structure/UnholdBankMoneyTransactionProcessorAgent.java
	BNK/plugin/bank_money_transaction/fermat-bnk-plugin-bank-money-transaction-withdraw-bitdubai/src/main/java/com/bitdubai/fermat_bnk_plugin/layer/bank_money_transaction/withdraw/developer/bitdubai/version_1/structure/BankMoneyTransactionRecordImpl.java
	BNK/plugin/bank_money_transaction/fermat-bnk-plugin-bank-money-transaction-withdraw-bitdubai/src/main/java/com/bitdubai/fermat_bnk_plugin/layer/bank_money_transaction/withdraw/developer/bitdubai/version_1/structure/WithdrawBankMoneyTransactionManager.java
	BNK/plugin/wallet/fermat-bnk-plugin-wallet-bank-money-bitdubai/src/main/java/com/bitdubai/fermat_bnk_plugin/layer/wallet/bank_money/developer/bitdubai/version_1/database/BankMoneyWalletDao.java
	BNK/plugin/wallet/fermat-bnk-plugin-wallet-bank-money-bitdubai/src/main/java/com/bitdubai/fermat_bnk_plugin/layer/wallet/bank_money/developer/bitdubai/version_1/structure/BankMoneyTransactionRecordImpl.java
	CBP/android/reference_wallet/fermat-cbp-android-reference-wallet-crypto-broker-bitdubai/src/main/java/com/bitdubai/reference_wallet/crypto_broker_wallet/fragments/settings/SetttingsStockManagementFragment.java
	CSH/android/reference_wallet/fermat-csh-android-reference-wallet-cash-money-bitdubai/src/main/java/com/bitdubai/reference_wallet/cash_money_wallet/fragments/transactionDetail/TransactionDetailFragment.java

+++ BNK/plugin/bank_money_transaction/fermat-bnk-plugin-bank-money-transaction-deposit-bitdubai/src/main/java/com/bitdubai/fermat_bnk_plugin/layer/bank_money_transaction/deposit/developer/bitdubai/version_1/structure/BankMoneyTransactionRecordImpl.java
@@ -10,6 +10,7 @@ import com.bitdubai.fermat_bnk_api.all_definition.enums.TransactionType;
 import com.bitdubai.fermat_bnk_api.layer.bnk_wallet.bank_money.interfaces.BankMoneyTransactionRecord;
 import com.bitdubai.fermat_bnk_plugin.layer.bank_money_transaction.deposit.developer.bitdubai.version_1.DepositBankMoneyTransactionPluginRoot;
 
+import java.math.BigDecimal;
 import java.util.UUID;
 
 /**
@@ -22,15 +23,15 @@ public class BankMoneyTransactionRecordImpl implements BankMoneyTransactionRecor
     String publicKeyCustomer;
     String balanceType;
     String transactionType;
-    float amount;
+    BigDecimal amount;
     String cashCurrencyType;
     String bankOperationType;
     String bankDocumentReference;
     String bankName;
     String bankAccountNumber;
     String bankAccountType;
-    long runningBookBalance;
-    long runningAvailableBalance;
+    BigDecimal runningBookBalance;
+    BigDecimal runningAvailableBalance;
     long timeStamp;
     String memo;
     String status;
@@ -81,7 +82,7 @@ public class BankMoneyTransactionRecordImpl implements BankMoneyTransactionRecor
     }
 
     @Override
-    public float getAmount() {
+    public BigDecimal getAmount() {
         return amount;
     }
 
@@ -136,12 +137,12 @@ public class BankMoneyTransactionRecordImpl implements BankMoneyTransactionRecor
     }
 
     @Override
-    public long getRunningBookBalance() {
+    public BigDecimal getRunningBookBalance() {
         return runningBookBalance;
     }
 
     @Override
-    public long getRunningAvailableBalance() {
+    public BigDecimal getRunningAvailableBalance() {
         return runningAvailableBalance;
     }
