+++ CBP/library/api/fermat-cbp-api/src/main/java/com/bitdubai/fermat_cbp_api/layer/wallet_module/crypto_broker/interfaces/CryptoBrokerWalletModuleManager.java
@@ -673,9 +673,9 @@ public interface CryptoBrokerWalletModuleManager extends CBPWalletsModuleManager
     /**
      * Returns the Balance this BankMoneyWalletBalance belongs to. (Can be available or book)
      *
-     * @return A double, containing the balance.
+     * @return A BigDecimal, containing the balance.
      */
-    double getBalanceBankWallet(String walletPublicKey, String accountNumber)
+    BigDecimal getBalanceBankWallet(String walletPublicKey, String accountNumber)
             throws CantCalculateBalanceException, CantLoadBankMoneyWalletException;
 
     /**
